introduce pip package, rename folders for unification
