+++ README.md
@@ -57,7 +57,7 @@ my_dessca_instance0.plot_scatter()
 
 Output:
 
-![](Figures/Scatter0.png)
+![](figures/Scatter0.png)
 
 And a corresponding coverage heatmap
 
@@ -68,7 +68,7 @@ my_dessca_instance0.plot_heatmap()
 
 Output:
 
-![](Figures/Heatmap0.png)
+![](figures/Heatmap0.png)
 
 The coverage probability density function (PDF) is updated with the given distribution.
 DESSCA can now suggest where to place the next sample.
@@ -92,7 +92,7 @@ my_dessca_instance0.plot_scatter()
 
 Output:
 
-![](Figures/Scatter1.png)
+![](figures/Scatter1.png)
 
 Let's have a look at the density:
 
@@ -102,7 +102,7 @@ my_dessca_instance0.plot_heatmap()
 
 Output:
 
-![](Figures/Heatmap1.png)
+![](figures/Heatmap1.png)
 
 ### More Features
 The scatter plots can also be rendered in an online fashion (100 samples):
@@ -120,7 +120,7 @@ for _ in range(100):
 
 Output:
 
-![](Figures/DESSCA_default.gif)
+![](figures/DESSCA_default.gif)
 
 Further, we can parameterize a memory buffer to only memorize a limited number of past samples:
 
@@ -139,7 +139,7 @@ for _ in range(100):
 
 Output:
 
-![](Figures/DESSCA_buffer.gif)
+![](figures/DESSCA_buffer.gif)
 
 See how forgetting past samples leads to a group of samples in a similar area?
 Lastly, we can also choose to use a specific reference coverage density:
@@ -165,7 +165,7 @@ for _ in range(100):
 
 Output:
 
-![](Figures/DESSCA_reference.gif)
+![](figures/DESSCA_reference.gif)
 
 DESSCA can also be used for downsampling. Let's firstly find a large dataset that we would like to reduce in size:
 
@@ -184,7 +184,7 @@ my_dessca_instance4.plot_scatter()
 
 Output:
 
-![](Figures/Scatter2.png)
+![](figures/Scatter2.png)
 
 Now, DESSCA can be used to reduce the set down to a specified number of remaining samples while trying to preserve the original distribution:
 
@@ -200,6 +200,6 @@ my_dessca_instance5.plot_scatter()
 
 Output:
 
-![](Figures/Scatter3.png)
+![](figures/Scatter3.png)
 
 Comparing the edge distributions, it can be seen that they are still almost the same despite removing 90 % of the dataset's content.
