+++ examples/PMSM_CurrentControl/Plot_TimeDomain.py
