+++ examples/PMSM_CurrentControl/ddpg_pmsm_dqCC_validation.py
@@ -1,8 +1,6 @@
 from tensorflow.keras.models import Sequential, Model
 from tensorflow.keras.layers import Dense, Flatten, Input, \
     Concatenate, LeakyReLU
-from tensorflow.keras import initializers, regularizers
-import tensorflow as tf
 from tensorflow.keras.optimizers import Adam
 from rl.agents import DDPGAgent
 from rl.memory import SequentialMemory
@@ -13,22 +11,16 @@ import sys
 import os
 sys.path.append(os.path.abspath(os.path.join('..')))
 import gym_electric_motor as gem
-from gym_electric_motor.reference_generators import MultipleReferenceGenerator, ConstReferenceGenerator, \
-    WienerProcessReferenceGenerator
-from gym_electric_motor.visualization import MotorDashboard
-from gym_electric_motor.physical_systems import ConstantSpeedLoad
+from gym_electric_motor.reference_generators import MultipleReferenceGenerator, ConstReferenceGenerator
+from gym_electric_motor.physical_systems import ExternalSpeedLoad
 from gym.core import Wrapper
 from gym.spaces import Box, Tuple
 import h5py
 from pathlib import Path
-import os, sys, inspect
-sys.path.append("../..")
-from DESSCA import dessca_model
-
 from multiprocessing import Pool
 
 
-use_dessca = False
+use_dessca = True
 
 
 # globals:
@@ -41,48 +33,86 @@ l_d = 0.37e-3
 l_q = 1.2e-3
 p = 3
 omega_lim = 12e3 * 2 * np.pi / 60
+tau = 1e-4
 
-memory_buffer_size = 50000
+rng = np.random.default_rng(42)
+i_step = i_n / 8
+i_dq_grid = np.array([[], []])
+for d_idx in range(9):
+    for q_idx in range(17):
+        i_d_new = - d_idx * i_step * 0.999
+        i_q_new = (q_idx - 8) * i_step * 0.999
+        if np.sqrt(i_d_new ** 2 + i_q_new ** 2) < i_n:
+            i_dq_grid = np.append(i_dq_grid, np.array([[i_d_new], [i_q_new]]), axis=1)
+rng.shuffle(i_dq_grid, axis=1)
 
-def _ref_pdf(X):
-    # states that must be available:
-    # i_d, i_q, omega, cos_epsilon, sin_epsilon, u_d-1, u_q-1, i_d^*, i_q^*
-    # actual states: i_d, i_q, omega, epsilon, i_d^*, i_q^*
-    i_d = X[0] * i_lim
-    i_q = X[1] * i_lim
+step_length_per_point = 300
+grid_test_duration_steps = np.shape(i_dq_grid)[1] * step_length_per_point
+grid_test_duration_time = grid_test_duration_steps * tau
+acceleration_steps = 5000
+accel_time = acceleration_steps * tau
 
-    omega = X[2] * omega_lim # box constraints
-    #epsilon = X[3] * np.pi # box constraints
-    i_d_ref = X[4] * i_lim
-    i_q_ref = X[5] * i_lim
+last_state_hold_steps = 5000
 
-    # plant
-    # i_d
-    _i_d_upper = np.less(i_d, np.clip((- psi_p + U_dc / (np.sqrt(3) * p * np.abs(omega))) / l_d, None, 0))
-    _i_d_lower = np.greater(i_d, np.clip((- psi_p - U_dc / (np.sqrt(3) * p * np.abs(omega))) / l_d, - i_n, None))
-    _i_d_allow = np.logical_and(_i_d_lower, _i_d_upper)
-    # i_q
-    _i_q_c_max = np.sqrt((U_dc / (l_d * np.sqrt(3) * p * np.abs(omega))) ** 2 - ((l_d * i_d + psi_p) / l_q) ** 2)
-    _i_q_upper = np.less(i_q, np.minimum(_i_q_c_max, np.sqrt(np.clip(i_n ** 2 - i_d ** 2, 0, None))))
-    _i_q_lower = np.greater(i_q, np.maximum(- _i_q_c_max, -np.sqrt(np.clip(i_n ** 2 - i_d ** 2, 0, None))))
-    _i_q_allow = np.logical_and(_i_q_lower, _i_q_upper)
-    _plant_allow = np.logical_and(_i_d_allow, _i_q_allow).astype(float)
+test_duration = (grid_test_duration_steps + acceleration_steps) * 5 + last_state_hold_steps
+max_return = test_duration * 0.1
+print(max_return)
+print()
+print()
+print()
 
-    # reference
-    # i_d_ref
-    _i_d_upper = np.less(i_d_ref, np.clip((- psi_p + U_dc / (np.sqrt(3) * p * np.abs(omega))) / l_d, None, 0))
-    _i_d_lower = np.greater(i_d_ref, np.clip((- psi_p - U_dc / (np.sqrt(3) * p * np.abs(omega))) / l_d, -i_n, None))
-    _i_d_ref_allow = np.logical_and(_i_d_lower, _i_d_upper)
-    # i_q_ref
-    _i_q_c_max = np.sqrt((U_dc / (l_d * np.sqrt(3) * p * np.abs(omega))) ** 2 - ((l_d * i_d_ref + psi_p) / l_q) ** 2)
-    _i_q_upper = np.less(i_q_ref, np.minimum(_i_q_c_max, np.sqrt(np.clip(i_n ** 2 - i_d_ref ** 2, 0, None))))
-    _i_q_lower = np.greater(i_q_ref, np.maximum(- _i_q_c_max, -np.sqrt(np.clip(i_n ** 2 - i_d_ref ** 2, 0, None))))
-    _i_q_ref_allow = np.logical_and(_i_q_lower, _i_q_upper)
-    _reference_allow = np.logical_and(_i_d_ref_allow, _i_q_ref_allow)
+def i_dq_validation_profile(k):
 
-    _state_filter = np.logical_and(_plant_allow, _reference_allow).astype(float)
+    _k = k % (grid_test_duration_steps + acceleration_steps)
+    if k >= test_duration - last_state_hold_steps:
+        i_d_ref = 0
+        i_q_ref = 0
+    elif _k < grid_test_duration_steps:
+        _point_idx = _k // step_length_per_point
+        i_d_ref = i_dq_grid[0, _point_idx] / i_lim
+        i_q_ref = i_dq_grid[1, _point_idx] / i_lim
+    else:
+        i_d_ref = 0
+        i_q_ref = 0
 
-    return _state_filter
+    return i_d_ref, i_q_ref
+
+def speed_profile(t):
+
+    niveau0 = 0.0
+    niveau1 = 0.15
+    niveau2 = 0.5
+
+    if t < grid_test_duration_time:
+        omega = niveau0 * omega_lim
+    elif t < (grid_test_duration_time + accel_time):
+        omega = ((t - grid_test_duration_time) * (niveau1 - niveau0) / accel_time + niveau0) * omega_lim
+
+    elif t < 2 * grid_test_duration_time + accel_time:
+        omega = niveau1 * omega_lim
+    elif t < 2 * (grid_test_duration_time + accel_time):
+        omega = ((t - 2 * grid_test_duration_time - accel_time) * (- niveau1 - niveau1) / accel_time + niveau1) * omega_lim
+
+    elif t < 3 * grid_test_duration_time + 2 * accel_time:
+        omega = - niveau1 * omega_lim
+    elif t < 3 * (grid_test_duration_time + accel_time):
+        omega = ((t - 3 * grid_test_duration_time - 2 * accel_time) * (niveau2 + niveau1) / accel_time - niveau1) * omega_lim
+
+    elif t < 4 * grid_test_duration_time + 3 * accel_time:
+        omega = niveau2 * omega_lim
+    elif t < 4 * (grid_test_duration_time + accel_time):
+        omega = ((t - 4 * grid_test_duration_time - 3 * accel_time) * (- niveau2 - niveau2) / accel_time + niveau2) * omega_lim
+
+    elif t < 5 * grid_test_duration_time + 4 * accel_time:
+        omega = - niveau2 * omega_lim
+    elif t < 5 * (grid_test_duration_time + accel_time):
+        omega = ((t - 5 * grid_test_duration_time - 4 * accel_time) * (niveau0 + niveau2) / accel_time - niveau2) * omega_lim
+
+    else:
+        omega = 0.0
+
+    return omega
+
 
 class AppendLastActionWrapper(Wrapper):
 
@@ -103,6 +133,8 @@ class AppendLastActionWrapper(Wrapper):
         self.REWARD = []
         self.HISTORY = []
 
+        self.time_idx = 0
+
         self.episode_count = 0
         self.agent_idx = agent_idx
         if use_dessca:
@@ -110,35 +142,16 @@ class AppendLastActionWrapper(Wrapper):
         else:
             self.folder_name = "Uniform"
 
-        if use_dessca:
-            state_constraints = [[-1, 1],
-                                 [-1, 1],
-                                 [-1, 1],
-                                 [-1, 1],
-                                 [-1, 1],
-                                 [-1, 1]]
-            self.dessca_model = dessca_model(box_constraints=state_constraints,
-                                             reference_pdf=_ref_pdf,
-                                             state_names=["i_d",
-                                                          "i_q",
-                                                          "omega",
-                                                          "epsilon",
-                                                          "i_d^*",
-                                                          "i_q^*"],
-                                             buffer_size=memory_buffer_size)
-
     def step(self, action):
 
         (state, ref), rew, term, info = self.env.step(action)
-        state[2] += np.random.normal(0, 0.0001) # measurement noise
-        ref[0] += np.random.normal(0, 0.01)
-        ref[1] += np.random.normal(0, 0.01)
+        self.time_idx += 1
+
+        ref[0], ref[1] = i_dq_validation_profile(self.time_idx)
         state = np.concatenate((state[0:3], [np.cos(state[3] * np.pi), np.sin(state[3] * np.pi)], action))
 
         i_d = state[0]
         i_q = state[1]
-        omega = state[2]
-        epsilon = np.arctan2(state[4], state[3]) / np.pi
         i_d_ref = ref[0]
         i_q_ref = ref[1]
         r_d = (np.sqrt(np.abs(i_d_ref - i_d) / 2) + ((i_d_ref - i_d) / 2) ** 2) / 2
@@ -150,9 +163,6 @@ class AppendLastActionWrapper(Wrapper):
         else:
             rew = (2 - r_d - r_q) / 2 * (1 - 0.9)
 
-        if use_dessca:
-            self.dessca_model.update_coverage_pdf(data=np.transpose([i_d, i_q, omega, epsilon, i_d_ref, i_q_ref]))
-
         if term:
             rew = -1
 
@@ -161,6 +171,9 @@ class AppendLastActionWrapper(Wrapper):
 
         state[3] *= 0.1
         state[4] *= 0.1
+
+        if self.time_idx == test_duration:
+            self.reset()
 
         return (state, ref), rew, term, info
 
@@ -173,7 +186,7 @@ class AppendLastActionWrapper(Wrapper):
                     self.folder_name
                     + "/" + self.folder_name
                     + "_" + str(self.agent_idx)
-                    + "/" + "training"
+                    + "/" + "validation"
                     + "_" + str(self.episode_count)
                     + ".hdf5", "w") as f:
                 lim = f.create_dataset("limits", data=np.concatenate((self.limits[0:3], [1, 1],
@@ -188,37 +201,12 @@ class AppendLastActionWrapper(Wrapper):
 
         state, ref = self.env.reset()
 
-        if not use_dessca:
-            eps_0 = np.random.uniform(-1, 1) * np.pi
-            omega_0 = np.random.uniform(-1, 1) * self.env.physical_system.limits[0]
-            psi_p = self.env.physical_system.electrical_motor.motor_parameter["psi_p"]
-            l_d = self.env.physical_system.electrical_motor.motor_parameter["l_d"]
-            l_q = self.env.physical_system.electrical_motor.motor_parameter["l_q"]
-            p = self.env.physical_system.electrical_motor.motor_parameter["p"]
-            u_dc = self.env.physical_system.limits[-1]
-            dc_link_d = u_dc / (np.sqrt(3) * l_d * np.abs(omega_0 * p))
-            i_d_upper = np.clip(- psi_p / l_d + dc_link_d, None, 0)
-            i_d_lower = np.clip(- psi_p / l_d - dc_link_d, -i_n, None)
-            i_d_0 = np.random.uniform(i_d_lower, i_d_upper)
-            i_q_upper = np.clip(np.sqrt((u_dc / (np.sqrt(3) * omega_0 * p * l_q)) ** 2 -
-                                        (l_d / l_q * (i_d_0 + psi_p / l_d)) ** 2), None, np.sqrt(i_n ** 2 - i_d_0 ** 2))
-            i_q_lower = np.clip(- i_q_upper, - np.sqrt(i_n ** 2 - i_d_0 ** 2), None)
-            i_q_0 = np.random.uniform(i_q_lower, i_q_upper)
-
-            i_d_ref = np.random.uniform(i_d_lower, i_d_upper)
-            i_q_upper_ref = np.clip(np.sqrt((u_dc / (np.sqrt(3) * omega_0 * p * l_q)) ** 2 -
-                                        (l_d / l_q * (i_d_ref + psi_p / l_d)) ** 2), None, np.sqrt(i_n ** 2 - i_d_ref ** 2))
-            i_q_lower_ref = np.clip(- i_q_upper, - np.sqrt(i_n ** 2 - i_d_ref ** 2), None)
-            i_q_ref = np.random.uniform(i_q_lower_ref, i_q_upper_ref)
-        else:
-            # i_d, i_q, omega, epsilon, i_d^*, i_q^*
-            dessca_state = self.dessca_model.sample_optimally()
-            i_d_0 = dessca_state[0] * self.env.physical_system.limits[2]
-            i_q_0 = dessca_state[1] * self.env.physical_system.limits[2]
-            omega_0 = dessca_state[2] * self.env.physical_system.limits[0]
-            eps_0 = dessca_state[2] * np.pi
-            i_d_ref = dessca_state[4] * self.env.physical_system.limits[2]
-            i_q_ref = dessca_state[5] * self.env.physical_system.limits[2]
+        i_d_0 = 0
+        i_q_0 = 0
+        eps_0 = 0
+        omega_0 = 0
+        i_d_ref = 0
+        i_q_ref = 0
 
         self.env.physical_system._ode_solver.set_initial_value(np.array([omega_0, i_d_0, i_q_0, eps_0]))
         self.env.reference_generator._sub_generators[0]._reference_value = i_d_ref / self.env.physical_system.limits[2]
@@ -241,20 +229,18 @@ class AppendLastActionWrapper(Wrapper):
         state[3] *= 0.1
         state[4] *= 0.1
 
+        self.time_idx = 0
+
         return state, ref
 
 
-def train_agent(agent_idx):
-    tf.config.set_visible_devices([], 'GPU')
-    os.environ["CUDA_DEVICE_ORDER"] = "PCI_BUS_ID"
-    os.environ['CUDA_VISIBLE_DEVICES'] = '-1'
-
+def test_agent(agent_idx):
     d_generator = ConstReferenceGenerator('i_sd', 0)
     q_generator = ConstReferenceGenerator('i_sq', 0)
     rg = MultipleReferenceGenerator([d_generator, q_generator])
 
     motor_parameter = dict(
-        r_s=r_s, l_d=l_d, l_q=l_q, psi_p=psi_p, p=p, j_rotor=0.06
+        r_s=r_s, l_d=l_d, l_q=l_q, psi_p=psi_p, p=p, j_rotor=0.001
     )
 
     limit_values = dict(
@@ -267,14 +253,14 @@ def train_agent(agent_idx):
 
     env = gem.make(
         'PMSMCont-v1',
-        load=ConstantSpeedLoad(omega_fixed=1000 * np.pi / 30),
+        load=ExternalSpeedLoad(speed_profile=speed_profile, tau=tau),
         control_space='dq',
         ode_solver='scipy.solve_ivp', solver_kwargs={},
         reference_generator=rg,
         reward_weights={'i_sq': 0.5, 'i_sd': 0.5},
         reward_power=0.5,
-        observed_states=['i_sd', 'i_sq'],
-        tau=1e-4,
+        observed_states=None, # ['i_sd', 'i_sq'],
+        tau=tau,
         dead_time=True,
         u_sup=U_dc,
         motor_parameter=motor_parameter,
@@ -320,11 +306,10 @@ def train_agent(agent_idx):
     x = LeakyReLU(alpha=0.1)(x)
     x = Dense(1, activation='linear')(x)
     critic = Model(inputs=(action_input, observation_input), outputs=x)
-    print(critic.summary())
 
     # Define a memory buffer for the agent, allows to learn from past experiences
     memory = SequentialMemory(
-        limit=memory_buffer_size,
+        limit=0,
         window_length=window_length
     )
 
@@ -333,9 +318,9 @@ def train_agent(agent_idx):
     random_process = OrnsteinUhlenbeckProcess(
         theta=10,
         mu=0.0,
-        sigma=1,
+        sigma=0,
         dt=env.physical_system.tau,
-        sigma_min=0.01,
+        sigma_min=0,
         n_steps_annealing=290000,
         size=2
     )
@@ -349,7 +334,6 @@ def train_agent(agent_idx):
         critic_action_input=action_input,
         memory=memory,
         random_process=random_process,
-
         # Define the overall training parameters
         nb_steps_warmup_actor=2000,
         nb_steps_warmup_critic=1000,
@@ -358,37 +342,32 @@ def train_agent(agent_idx):
         batch_size=16,
         memory_interval=1
     )
-
     # Compile the function approximators within the agent (making them ready for training)
     # Note that the DDPG agent uses two function approximators, hence we define two optimizers here
     agent.compile([Adam(lr=5e-6), Adam(lr=5e-4)])
-
-    # Start training for 75 k simulation steps
-    agent.fit(
-        env,
-        nb_steps=400000,
-        nb_max_start_steps=0,
-        nb_max_episode_steps=100,
-        visualize=True,
-        action_repetition=1,
-        verbose=2,
-        log_interval=10000,
-        callbacks=[],
-    )
 
     if use_dessca:
         folder_name = "Dessca"
     else:
         folder_name = "Uniform"
-
-    agent.save_weights(filepath=folder_name +
+    agent.load_weights(filepath=folder_name +
                                 "/" + folder_name +
                                 "_" + str(agent_idx) +
                                 "/" + folder_name +
-                                "_weights.hdf5",
-                       overwrite=True)
+                                "_weights.hdf5"
+                       )
 
+    agent.test(
+        env,
+        nb_max_start_steps=0,
+        nb_max_episode_steps=test_duration,
+        nb_episodes=1,
+        visualize=False,
+        action_repetition=1,
+        verbose=2,
+        callbacks=[],
+    )
 
 if __name__ == '__main__':
     with Pool() as p:
-        p.map(train_agent, range(50))
+        p.map(test_agent, range(50))
