+++ pyproject.toml
@@ -1,0 +1,88 @@
+[project]
+name = "dessca"
+version = "0.1.0"
+authors = [
+    { name = "Maximilian Schenke" },
+]
+description = "Improved Exploring Starts by Kernel Density Estimation-Based State-Space Coverage Acceleration in Reinforcement Learning."
+readme = "README.md"
+requires-python = ">=3.11"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+dynamic = ["dependencies"] # commented due to packaging issues: "optional-dependencies"
+
+[tool.setuptools.dynamic]
+dependencies = { file = ["requirements.txt"] }
+optional-dependencies = { dev = { file = ["requirements-dev.txt"] } }
+
+[project.urls]
+Homepage = "https://github.com/max-schenke/DESSCA"
+Issues = "https://github.com/max-schenke/DESSCA/issues"
+
+[build-system]
+requires = ["hatchling", "hatch-requirements-txt"]
+build-backend = "hatchling.build"
+
+[tool.hatch.metadata.hooks.requirements_txt]
+files = ["requirements.txt"]
+
+[tool.hatch.build.targets.wheel]
+packages = ["dessca/"]
+
+[tool.hatch.build.targets.sdist]
+include = [
+    "dessca/*.py",
+    "requirements.txt"
+]
+
+
+
+[tool.ruff]
+exclude = [
+    ".eggs",
+    ".git",
+    ".venv",
+    "venv"]
+
+line-length = 120
+indent-width = 4
+
+target-version = "py311"
+
+[tool.ruff.lint]
+select = ["E4", "E7", "E9", "F", "B", "D", "D417"]
+# extend-select = ["D417"] deactivated by default in case of pep257 codestyle.
+extend-select = ["PLR0203", "PLW0211", "PLR6301"]
+preview = true
+explicit-preview-rules = true
+# see also: https://docs.astral.sh/ruff/rules/undocumented-param/
+ignore = ["B008", "D107", "D203", "D212", "D213", "D402", "D413", "D415", "D416", "E722", "E731", "F403", "F405", "F841",]
+fixable = ["ALL"]
+unfixable = []
+# ignore list in docstring according to numpy codestyles for Dxxx.
+# http://www.pydocstyle.org/en/5.0.1/error_codes.html#default-conventions
+
+
+[tool.ruff.lint.pydocstyle]
+convention = "pep257"
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
+skip-magic-trailing-comma = false
+line-ending = "auto"
+
+
+
+[tool.mypy]
+python_version = "3.11"
+warn_return_any = true
+warn_unused_configs = true
+disallow_incomplete_defs = true
+exclude = [
+]
+
+ignore_missing_imports = true
